+++ optimizer/tests/test_notification.py
@@ -14,7 +14,7 @@ client = TestClient(app)
 # ---------------------------------------------------------------------------
 
 class TestChatReminderEndpoints:
-    @patch("optimizer.api.routes.send_chat_dms")
+    @patch("optimizer.api.routes_notify.send_chat_dms")
     def test_chat_reminder_success(self, mock_send: MagicMock) -> None:
         """POST /notify/chat-reminder → 200"""
         mock_send.return_value = (1, [
@@ -37,7 +37,7 @@ class TestChatReminderEndpoints:
         assert data["results"][0]["staff_id"] == "h1"
         assert data["results"][0]["success"] is True
 
-    @patch("optimizer.api.routes.send_chat_dms")
+    @patch("optimizer.api.routes_notify.send_chat_dms")
     def test_chat_reminder_partial_failure(self, mock_send: MagicMock) -> None:
         """一部失敗でも200を返し結果に反映される"""
         mock_send.return_value = (1, [
@@ -71,7 +71,7 @@ class TestChatReminderEndpoints:
         )
         assert response.status_code == 422
 
-    @patch("optimizer.api.routes.send_chat_dms")
+    @patch("optimizer.api.routes_notify.send_chat_dms")
     def test_chat_reminder_custom_message(self, mock_send: MagicMock) -> None:
         """カスタムメッセージが send_chat_dms に渡される"""
         mock_send.return_value = (1, [
